Mapper クラスを作成し、json 関連処理をまとめる
