+++ ios-training-bjung/Models/WeatherModel.swift
@@ -15,27 +15,7 @@ protocol WeatherModelInput {
 }
 
 final class WeatherModel: WeatherModelInput {
-    
-    static let dateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
-        return dateFormatter
-    }()
-    
-    static let encoder: JSONEncoder = {
-        let encoder = JSONEncoder()
-        encoder.keyEncodingStrategy = .convertToSnakeCase
-        encoder.dateEncodingStrategy = .formatted(dateFormatter)
-        return encoder
-    }()
-    
-    static let decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        decoder.dateDecodingStrategy = .formatted(dateFormatter)
-        return decoder
-    }()
-    
+
     private var yumemiWeather: YumemiWeather.Type
     
     init(yumemiWeather: YumemiWeather.Type) {
@@ -51,15 +31,10 @@ final class WeatherModel: WeatherModelInput {
     }
     
     func fetchWeatherData(request: WeatherRequest) async throws -> WeatherResponse {
-        let jsonData = try WeatherModel.encoder.encode(request)
-        guard let jsonString = String(data: jsonData, encoding: .utf8) else { fatalError("Failed to encode to JSON") }
+        let mapper = Mapper.shared
+        let jsonString = try mapper.encodeWeatherRequest(request: request)
         let response = try yumemiWeather.fetchWeather(jsonString)
-        let decoder = WeatherModel.decoder
-        guard let responseData = response.data(using: .utf8),
-              let result = try? decoder.decode(WeatherResponse.self, from: responseData) else {
-            fatalError("Failed to encode to JSON")
-        }
-        return result
+        return try mapper.decodeWeatherResponse(json: response)
     }
 }
